FIR/UAST: commonize base UClass (with bogus annotation abstraction)
