+++ plugins/uast-kotlin-fir/testData/legacyRenderLog/AnnotationComplex.log.fir.txt
@@ -2,16 +2,21 @@ UFile (package = )
     UClass (name = Annotation)
         UMethod (name = strings)
     UClass (name = A)
+        UAnnotation (fqName = not-implemented-annotation)
         UMethod (name = A)
     UClass (name = AnnotationInner)
         UMethod (name = value)
     UClass (name = B1)
+        UAnnotation (fqName = not-implemented-annotation)
         UMethod (name = B1)
     UClass (name = B2)
+        UAnnotation (fqName = not-implemented-annotation)
         UMethod (name = B2)
     UClass (name = AnnotationArray)
         UMethod (name = value)
     UClass (name = C)
+        UAnnotation (fqName = not-implemented-annotation)
         UMethod (name = C)
     UClass (name = C2)
+        UAnnotation (fqName = not-implemented-annotation)
         UMethod (name = C2)
