+++ plugins/uast-kotlin-fir/testData/legacyRenderLog/ClassAnnotation.log.fir.txt
@@ -1,15 +1,20 @@
 UFile (package = )
     UClass (name = A)
+        UAnnotation (fqName = not-implemented-annotation)
         UMethod (name = A)
     UClass (name = MyAnnotation)
         UMethod (name = text)
     UClass (name = B)
+        UAnnotation (fqName = not-implemented-annotation)
         UField (name = Companion)
         UMethod (name = B)
         UClass (name = InB)
+            UAnnotation (fqName = not-implemented-annotation)
             UMethod (name = InB)
         UClass (name = Companion)
+            UAnnotation (fqName = not-implemented-annotation)
             UMethod (name = Companion)
     UClass (name = Obj)
+        UAnnotation (fqName = not-implemented-annotation)
         UField (name = INSTANCE)
         UMethod (name = Obj)
