+++ plugins/uast-kotlin-fir/testData/legacyTypes/ClassAnnotation.types.fir.txt
@@ -1,15 +1,20 @@
 UFile (package = ) [public final class A {...]
     UClass (name = A) [public final class A {...}]
+        UAnnotation (fqName = not-implemented-annotation) [@not-implemented-annotation]
         UMethod (name = A) [public fun A() = UastEmptyExpression]
     UClass (name = MyAnnotation) [public abstract annotation MyAnnotation {...}]
         UMethod (name = text) [public abstract fun text() : java.lang.String = UastEmptyExpression]
     UClass (name = B) [public final class B {...}]
+        UAnnotation (fqName = not-implemented-annotation) [@not-implemented-annotation]
         UField (name = Companion) [public static final var Companion: B.Companion]
         UMethod (name = B) [public fun B() = UastEmptyExpression]
         UClass (name = InB) [public static final class InB {...}]
+            UAnnotation (fqName = not-implemented-annotation) [@not-implemented-annotation]
             UMethod (name = InB) [public fun InB() = UastEmptyExpression]
         UClass (name = Companion) [public static final class Companion {...}]
+            UAnnotation (fqName = not-implemented-annotation) [@not-implemented-annotation]
             UMethod (name = Companion) [private fun Companion() = UastEmptyExpression]
     UClass (name = Obj) [public final class Obj {...}]
+        UAnnotation (fqName = not-implemented-annotation) [@not-implemented-annotation]
         UField (name = INSTANCE) [public static final var INSTANCE: Obj]
         UMethod (name = Obj) [private fun Obj() = UastEmptyExpression]
